Convert all offset/length values to int64
The POSIX standard specifies that these values should be represented
by a signed integer, and the Go stdlib adheres to this standard. The
llapi package in go-lustre handles safely converting from uint64
to int64 so that the rest of our code is simpler.

+++ cmd/ldmc/hsm.go
@@ -200,7 +200,7 @@ func getPathStatus(c *cli.Context, filePath string) (string, error) {
 					return "", errors.Wrapf(err, "Failed to stat() %s", filePath)
 				}
 				fmt.Fprintf(&buf, "(%s/%s)",
-					humanize.IBytes(a.BytesCopied),
+					humanize.IBytes(uint64(a.BytesCopied)),
 					humanize.IBytes(uint64(st.Size())))
 			}
 		}
